+++ R/create_any_autoimmune.R
@@ -14,15 +14,9 @@ create_any_autoimmune = function(
   dataset |>
     mutate(
       "any autoimmune disorder" = case_when(
-        dplyr::if_any(
-          .cols = all_of(autoimmune_dz),
-          .fns = ~ . %in% "Yes"
-        ) ~ "Some autoimmune recorded",
-        dplyr::if_all(
-          .cols = all_of(autoimmune_dz),
-          .fns = ~ is.na(.)
-        ) ~ NA,
-        TRUE ~ "No autoimmune recorded"
+        dplyr::if_any(.cols = all_of(autoimmune_dz), .fns = ~ . %in% "Yes") ~ TRUE,
+        dplyr::if_all(.cols = all_of(autoimmune_dz), .fns = ~ is.na(.)) ~ NA,
+        TRUE ~ FALSE
       ) |>
         factor() |>
         relevel(ref = "No autoimmune recorded")