+++ NEWS.md
@@ -1,7 +1,7 @@
 # fxtas (development version)
 
 * fixed `create_any_tremor()` to handle NAs before or after `fix_factors()`.
-* renamed "Any Autoimmune" to "any autoimmune disorder"
+* renamed "Any Autoimmune" to "any autoimmune disorder" and make it boolean
 * decapitalized "MMSE Total Score" to "MMSE total score"
 * decapitalized "Head tremor" to "head tremor"
 * decapitalized "Parkinsonian features" to "parkinsonian features"